requirements.txt
asyncpg
discord.py
# jishaku is optional
jishaku 